+++ pushapkscript/src/mozapkpublisher/test/test_push_apk.py
@@ -24,9 +24,6 @@ apk_x86 = NamedTemporaryFile()
 apk_arm = NamedTemporaryFile()
 
 APKS = [apk_x86, apk_arm]
-SERVICE_ACCOUNT = 'foo@developer.gserviceaccount.com'
-CLIENT_ID = 'client'
-CLIENT_SECRET = 'secret'
 
 
 def patch_extract_metadata(monkeypatch):
@@ -78,7 +75,7 @@ def patch_store_transaction(monkeypatch_, patch_target):
     mock_edit = create_autospec(patch_target)
 
     @contextmanager
-    def fake_transaction(_, __, ___, *, contact_server, dry_run):
+    def fake_transaction(_, __, *, contact_server, dry_run):
         yield mock_edit
 
     monkeypatch_.setattr(patch_target, 'transaction', fake_transaction)
@@ -88,7 +85,7 @@ def patch_store_transaction(monkeypatch_, patch_target):
 def test_google(monkeypatch):
     mock_metadata = patch_extract_metadata(monkeypatch)
     edit_mock = patch_store_transaction(monkeypatch, store.GooglePlayEdit)
-    push_apk(APKS, SERVICE_ACCOUNT, credentials, [], 'rollout', rollout_percentage=50,
+    push_apk(APKS, credentials, [], 'rollout', rollout_percentage=50,
              contact_server=False)
     edit_mock.update_app.assert_called_once_with([
         (apk_arm, mock_metadata[apk_arm]),
@@ -102,7 +99,7 @@ def test_push_apk_tunes_down_logs(monkeypatch):
     monkeypatch.setattr('mozapkpublisher.push_apk.extract_and_check_apks_metadata', MagicMock())
     monkeypatch.setattr('mozapkpublisher.common.utils.metadata_by_package_name', MagicMock())
 
-    push_apk(APKS, SERVICE_ACCOUNT, credentials, [], 'alpha', contact_server=False)
+    push_apk(APKS, credentials, [], 'alpha', contact_server=False)
 
     main_logging_mock.init.assert_called_once_with()
 
@@ -118,7 +115,6 @@ def test_main_google(monkeypatch):
     file = os.path.join(os.path.dirname(__file__), 'data', 'blob')
     fail_manual_validation_args = [
         'script',
-        '--username', 'foo@developer.gserviceaccount.com',
         '--secret', file,
         'alpha',
         file,
@@ -131,7 +127,6 @@ def test_main_google(monkeypatch):
 
         mock_push_apk.assert_called_once_with(
             ANY,
-            'foo@developer.gserviceaccount.com',
             file,
             ['org.mozilla.fennec_aurora'],
             'alpha',