+++ pushapkscript/src/mozapkpublisher/CHANGELOG.md
@@ -2,10 +2,15 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [7.0.0] - 2023-XX-XX
+
+### Changed
+* Switched Google Play Store authentication to the google-auth library.  The credentials now need to be passed as a json file instead of an email address and PKCS#12 file.
+
 ## [6.3.0] - 2023-11-08
 
 ### Removed
-* Amazon Store support
+* Removed support for pushing to the Amazon store.
 
 ### Fixed
 * Better diagnostics during AAB upload